# backend/docs/Swagger Documentation Package/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Prospera API",
    "version": "1.0.0",
    "description": "API Documentation for Prospera Backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "Menampilkan Seluruh Produk",
        "description": "Mengambil semua produk milik toko dengan dukungan paginasi, status stok, dan saran jumlah restock pintar (AI).\n\n**Otorisasi (RBAC):** Owner | Karyawan\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Nomor halaman",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Jumlah item per halaman",
            "schema": {
              "type": "integer",
              "default": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar produk berhasil diambil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalItems": {
                      "type": "integer",
                      "example": 45
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 5
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Product"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "stock_status": {
                                "type": "string",
                                "example": "AMAN"
                              },
                              "suggested_restock": {
                                "type": "integer",
                                "example": 12
                              },
                              "velocity": {
                                "type": "number",
                                "example": 0.45
                              },
                              "Category": {
                                "$ref": "#/components/schemas/Category"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Menambahkan Produk Baru",
        "description": "Menambahkan data produk baru ke dalam inventaris toko.\n\n**Otorisasi (RBAC):** Owner Only\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "product_name",
                  "product_cost",
                  "product_price",
                  "product_stock"
                ],
                "properties": {
                  "product_name": {
                    "type": "string",
                    "example": "Susu UHT Cokelat"
                  },
                  "product_cost": {
                    "type": "integer",
                    "example": 4500
                  },
                  "product_price": {
                    "type": "integer",
                    "example": 6000
                  },
                  "product_stock": {
                    "type": "integer",
                    "example": 50
                  },
                  "category_id_fk": {
                    "type": "integer",
                    "example": 1
                  },
                  "expired_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-12-31"
                  }
                }
              },
              "examples": {
                "createProduct": {
                  "$ref": "#/components/examples/CreateProductExample"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk berhasil ditambahkan.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Produk berhasil ditambahkan ke toko Anda."
                    },
                    "productId": {
                      "type": "integer",
                      "example": 12
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal (misal harga < 500, stok negatif, atau nama kembar).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Akses ditolak (hanya owner yang diizinkan).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/forecast": {
      "get": {
        "tags": [
          "Forecast"
        ],
        "summary": "Prediksi Penjualan Masa Depan (Forecasting)",
        "description": "Menghasilkan prediksi penjualan masa depan berdasarkan pola transaksi historis.\n\n**Otorisasi (RBAC):** Owner Only\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Hasil prediksi penjualan berhasil didapatkan.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ForecastResult"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Melihat Seluruh Kategori Produk",
        "description": "Menampilkan daftar seluruh kategori produk yang ada di toko.\n\n**Otorisasi (RBAC):** Owner | Karyawan\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar kategori berhasil diambil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Membuat Kategori Produk Baru",
        "description": "Menambahkan kategori baru ke dalam daftar kategori produk toko.\n\n**Otorisasi (RBAC):** Owner Only\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              },
              "example": {
                "category_name": "Minuman",
                "requires_expired_date": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kategori baru berhasil dibuat.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Kategori berhasil dibuat."
                    },
                    "category_id": {
                      "type": "integer",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/store-settings": {
      "get": {
        "tags": [
          "Store Settings"
        ],
        "summary": "Melihat Pengaturan Toko",
        "description": "Mengambil data konfigurasi/pengaturan toko saat ini.\n\n**Otorisasi (RBAC):** Owner | Karyawan\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pengaturan toko berhasil diambil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreSetting"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": [
          "Store Settings"
        ],
        "summary": "Memperbarui Pengaturan Toko",
        "description": "Mengubah konfigurasi jam buka, Grace Period, overtime, dsb.\n\n**Otorisasi (RBAC):** Owner Only\n**Keamanan:** Memerlukan JWT Token (Bearer Auth)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreSetting"
              },
              "example": {
                "open_hour": "08:00:00",
                "close_hour": "22:00:00",
                "grace_period_minutes": 30,
                "is_overtime_active": false,
                "emergency_pin": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pengaturan toko berhasil diperbarui.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Pengaturan toko berhasil disimpan."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "user_id_fk": {
            "type": "integer",
            "example": 1
          },
          "product_name": {
            "type": "string",
            "example": "Indomie Goreng"
          },
          "product_cost": {
            "type": "integer",
            "example": 2500
          },
          "product_price": {
            "type": "integer",
            "example": 3500
          },
          "product_stock": {
            "type": "integer",
            "example": 100
          },
          "category_id_fk": {
            "type": "integer",
            "example": 1
          },
          "expired_date": {
            "type": "string",
            "format": "date",
            "example": "2026-12-31"
          },
          "min_display_qty": {
            "type": "integer",
            "example": 5
          },
          "calculated_reorder_point": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "example": 1
          },
          "user_id_fk": {
            "type": "integer",
            "example": 1
          },
          "category_name": {
            "type": "string",
            "example": "Makanan"
          },
          "requires_expired_date": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "StoreSetting": {
        "type": "object",
        "properties": {
          "setting_id": {
            "type": "integer",
            "example": 1
          },
          "user_id_fk": {
            "type": "integer",
            "example": 1
          },
          "open_hour": {
            "type": "string",
            "example": "08:00:00"
          },
          "close_hour": {
            "type": "string",
            "example": "22:00:00"
          },
          "grace_period_minutes": {
            "type": "integer",
            "example": 30
          },
          "is_overtime_active": {
            "type": "boolean",
            "example": false
          },
          "emergency_pin": {
            "type": "string",
            "example": "hashed_pin_value_here"
          }
        }
      },
...
      "ForecastResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Prediksi penjualan harian (7 hari terakhir, tanpa outlier, hanya transaksi sukses)"
          },
          "prediction": {
            "type": "string",
            "example": "Rp 320000"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Terjadi kesalahan internal sistem yang tidak terduga."
          }
        }
      }
    },
    "examples": {
...
      "CreateProductExample": {
        "summary": "Create product example",
        "value": {
          "product_name": "Susu UHT Cokelat",
          "product_cost": 4500,
          "product_price": 6000,
          "product_stock": 50,
          "category_id_fk": 1,
          "expired_date": "2026-12-31"
        }
      },
...
    }
  }
}